requirements.txt
anyio==4.8.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.0
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.3.0
cleo==2.1.0
click==8.3.1
crashtest==0.4.1
cryptography==41.0.4
distlib==0.3.8
dulwich==0.24.2
fastjsonschema==2.18.1
filelock==3.16.1
findpython==0.6.2
Flask==3.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.4
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.3.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
Jinja2==3.1.6
keyring==25.6.0
MarkupSafe==3.0.3
more-itertools==10.1.0
msgpack==1.1.2
pbs-installer==2025.2.12
pillow==12.0.0
pipenv==2025.0.4
pkginfo==1.12.0
platformdirs==4.3.6
poetry==2.2.1
poetry-core==2.2.1
pycparser==2.21
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
RapidFuzz==3.14.1
reportlab==4.4.4
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.3
six==1.17.0
sniffio==1.3.1
tomlkit==0.12.1
trove-classifiers==2023.9.19
urllib3==2.5.0
virtualenv==20.35.4
Werkzeug==3.1.3
zstandard==0.23.0
